Move "lib" to "src"
