+++ src/augment-generator.js
